+++ packages/boethius-lang/CHANGELOG.md
@@ -3,6 +3,7 @@ v0.7.0
 - Curly bracket enclosed properties lists.
 - Identifiers at the beginning of scope can now be either builtin or scope property.
 - legato builtin.
+- Block comments
 
 v0.6.0
 ------